requirements.txt
streamlit
requests
python-dotenv
pypdf
python-docx